+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- Playlist: Space key doesn't create error when no sounds in playlist
+- Playlist: Sounds dragged out of top spot no longer stay yellow
+- Playlist: Search doc error fixed
+
 ## [0.3.1] - 2017-04-08
 ## Added
 - OS X build